refactor: create a PostCard component
